refactor: rename InstructionDecoder to ShaderMacroProcessor
